+++ lib/main.dart
@@ -5,6 +5,8 @@ import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:money/models/data_io/import_qfx.dart';
+import 'package:money/models/data_io/import_qif.dart';
 import 'package:money/models/settings.dart';
 import 'package:money/views/view_aliases.dart';
 import 'package:money/views/view_cashflow.dart';
@@ -107,8 +109,6 @@ class _MyMoneyState extends State<MyMoney> {
             settings.save();
             loadData();
           }
-        } else {
-          // todo: handle qfx, ofx, pdf, json, etc.
         }
       } catch (e) {
         debugLog(e.toString());
@@ -127,6 +127,21 @@ class _MyMoneyState extends State<MyMoney> {
     settings.pathToDatabase = Constants.demoData;
     settings.save();
     loadData();
+  }
+
+  void handleImport() async {
+    final FilePickerResult? pickerResult = await FilePicker.platform.pickFiles(type: FileType.any);
+    if (pickerResult != null) {
+      switch (pickerResult.files.single.extension?.toLowerCase()) {
+        case "qif":
+          importQIF(pickerResult.files.single.path.toString());
+        case "qfx":
+          importQFX(pickerResult.files.single.path.toString(), data);
+      }
+      setState(() {
+        // update UI
+      });
+    }
   }
 
   Widget showLoading() {
@@ -159,7 +174,13 @@ class _MyMoneyState extends State<MyMoney> {
 
   Widget welcomePanel(final BuildContext context) {
     return Scaffold(
-      appBar: createAppBar(settings, handleFileOpen, handleFileClose, onSettingsChanged),
+      appBar: createAppBar(
+        settings,
+        handleFileOpen,
+        handleFileClose,
+        handleImport,
+        onSettingsChanged,
+      ),
       body: Row(children: <Widget>[
         renderWelcomeAndOpen(context),
       ]),
@@ -253,7 +274,13 @@ class _MyMoneyState extends State<MyMoney> {
 
   Widget getScaffoldingForSmallSurface(final BuildContext context) {
     return Scaffold(
-      appBar: createAppBar(settings, handleFileOpen, handleFileClose, onSettingsChanged),
+      appBar: createAppBar(
+        settings,
+        handleFileOpen,
+        handleFileClose,
+        handleImport,
+        onSettingsChanged,
+      ),
       body: Row(children: <Widget>[Expanded(child: getWidgetForMainContent(context, settings.screenIndex))]),
       bottomNavigationBar:
           MenuHorizontal(settings: settings, onSelectItem: handleScreenChanged, selectedIndex: settings.screenIndex),
@@ -262,7 +289,13 @@ class _MyMoneyState extends State<MyMoney> {
 
   Widget getScaffoldingForLargeSurface(final BuildContext context) {
     return Scaffold(
-      appBar: createAppBar(settings, handleFileOpen, handleFileClose, onSettingsChanged),
+      appBar: createAppBar(
+        settings,
+        handleFileOpen,
+        handleFileClose,
+        handleImport,
+        onSettingsChanged,
+      ),
       body: SafeArea(
         bottom: false,
         top: false,